♻️ refactor(MediaProcessHandler): extraire MediaProcessor (SRP)
La logique de création du Media (résolution du type, idempotence,
EXIF, thumbnail) était figée dans le handler Messenger, donc
inaccessible sans passer par le bus async. Extraite dans
MediaProcessor::process(File): ?Media, derrière MediaProcessorInterface
(nécessaire pour le mock en test — la classe est final).

MediaProcessHandler devient un adaptateur fin : résout le File puis
délègue. Comportement strictement identique, pas de nouvelle feature
dans ce commit — prépare l'import direct dans un album (prochain
commit), qui a besoin du Media en synchrone juste après l'upload.

TDD : MediaProcessorTest (nouveau, couvre la logique extraite) +
MediaProcessHandlerTest (réécrit pour vérifier la délégation) —
RED → GREEN.
