+++ src/Interface/MediaProcessorInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Interface;
+
+use App\Entity\File;
+use App\Entity\Media;
+
+interface MediaProcessorInterface
+{
+    /**
+     * Crée (ou retourne, si déjà traité) le Media associé à un File.
+     * Retourne null si le type MIME du fichier n'est pas un média supporté.
+     */
+    public function process(File $file): ?Media;
+}
